+++ www/install/install.php
@@ -1,6 +1,10 @@
 <?php
 require('../vendor/autoload.php');
 
+ini_set('zlib.output_compression', 0);
+ini_set('implicit_flush', 1);
+
+
 require("gotest.php");
 if(!systemIsOk()) {
 	header("Location: test.php");
@@ -11,6 +15,8 @@ if(!systemIsOk()) {
 use GO\Base\Cron\CronJob;
 use GO\Base\Model\Module;
 use GO\Base\Observable;
+use go\modules\community\bookmarks\Module as BookmarksModule;
+use go\modules\community\comments\Module as CommentsModule;
 use go\core\App;
 use go\core\jmap\State;
 use go\core;
@@ -59,7 +65,9 @@ if (!empty($_POST)) {
 		App::get()->getInstaller()->install($admin, [
 				new AddressBookModule(), 
 				new NotesModule(),
-				new GAModule()
+				new GAModule(),
+				new CommentsModule(),
+				new BookmarksModule()
 				]);
 
 		//install not yet refactored modules