+++ www/views/Extjs3/css.php
@@ -1,9 +1,12 @@
 <?php
+
+use go\core\webclient\Extjs3;
+
 require(__DIR__ . "/../../vendor/autoload.php");
 
 \go\core\App::get();
 		
-$webclient = new \go\core\webclient\Extjs3();
+$webclient = Extjs3::get();
 $webclient->getCSSFile()->output(true, true, [
 	"Expires" => (new DateTime("1 year"))->format("D, j M Y H:i:s")
 ]);