+++ www/views/Extjs3/javascript/overrides.js
@@ -1017,6 +1017,10 @@ Ext.override(Ext.form.CompositeField, {
 	submit: false //don't submit with form.getFieldValue()
 });
 
+Ext.override(Ext.form.DisplayField, {
+	submit: false //don't submit with form.getFieldValue()
+});
+
 Ext.override(Ext.form.DateField, {
 	getValue : function(){
         return this.parseDate(Ext.form.DateField.superclass.getValue.call(this)) || null;